glob.c: Untested addition of jgm's fix for problems when LIST %.% occurred
reconstruct.c: When reconstruct encountered a MAILBOX_FORMAT_NETNEWS
mailbox that didn't have a corresponding directory in the newsspool,
it failed; now, it just assumes there are no articles.

configure.in: Changed "-R/usr/lib" to "-R /usr/lib" so now that
compilers that don't want -R and ignore it fail when they try to open
/usr/lib, and therefore don't get the -R switch when we get around to
compilation.