Accept 'From ' header from IMAP clients
Hotmail used to create a 'From ' header and export it via IMAP.
Outlook would send that header right up to the IMAP server.  Yet
another workaround for misbehaving clients.  The bane of our
existance they are, yes indeed.

This patch causes any line starting with 'From ' in the header
to be ignored rather than triggering a MESSAGE_BADHEADER error.