Clean Shutdown
The instructions for Cyrus say to send a signal to master to shut
the system down.  This generally means a SIGTERM.

When master gets a SIGTERM is sends SIGTERM to all children, who
don't trap it and exit instantly.  Fine if you're never in a
"critical section"

Unfortunately both replication and fast rename create critical
sections.  Fast rename thanks to split meta (two renames) and
replication because if you have successfully committed a message
to the index file but not yet logged the replication event, then
it could fail to be copied.  In a system as big as fastmail.fm's,
this means almost every shutdown causes a couple of messages to
fail to sync.  Then we bring up the replica, get two messages
with the same UID, hilarity ensues.

This patch adds a SIGQUIT handler to master, which sends SIGQUIT
to children then doesn't exit until all children have.  The
children already had a SIGQUIT handler which waited for
signals_poll() to be checked.  The patch also adds a few more
signals_poll() in likely places.

Some documentation has been added on using this, though I feel it
could be made a bit more prominent since clean shutdown is much
safer for data integrity.  In particular, distribution maintainers
should be making their system init scripts use it.