Add conversations_getpath()
which constructs and returns a new string which is the pathname
of the canonical location of the conversations db for a given
userid.  Uses strconcat() and mboxname_to_parts() to do this
cleanly (unlike every other path construction function in Cyrus).

Note that we deliberately don't combine this with conversations_open()
for two reasons:

a) it allows unit testing with non-canonical pathnames

b) it allows callers do to interesting things in between the
   two steps, which will become useful in a utility soon.