Do sync CRC calculations on-demand.
A mailbox' sync CRC is now calculated using a single full pass through
the mailbox, when either the sync client or sync server need to know it.
Some effects include:

 * The sync_crc field in the index header is now unused (I haven't
   bumped the version number or changed the on-disk format).

 * Incremental CRC updates are no longer performed, which means
   that the CRC algorithm used does not need to have this property.
   This frees us to use algorithms such as MD5 or SHA1 once the
   client and server have the ability to negotiate their use.

 * The only defined algorithm for now is the same CRC32 algorithm
   applied to the same fields, for backwards compatibility.

 * The value for SYNC_CRC in the dlists sent between the client and
   server is now an atom rather than a number.  For the current
   algorithm this has no effect on the wire format.

Note that backwards compatibility is preserved by this change.  The
format and semantics of the information on the wire do not change.
Older clients still work with newer servers and vice versa.