Parse server capabilities more carefully.
Instead of lots of strstr() and some blind luck, backend.c now carefully
parses the capability strings reported by the various servers it talks
to, and saves some of them for future reference.  In particular, the
parameter strings supplied with some capabilities are accumulated and
saved.

Because we now carefully parse the capabilities, some of the
pecularities of the various protocols now needed to be flagged in the
protocol_t structure so that backend.c can behave appropriately.

For example, most supported protocols report capabilities over multiple
lines, with each line containing one capability and any of its
parameters.  IMAP, on the other hand, reports them all in one big line
with parameters encoded as NAME=VALUE pairs.  Some protocols emit a
leading word in front of each of these lines; one doesn't.  Most
protocols emit the capability names and parameters as atoms; one
requires them to be quoted.

This work is necessary to allow capabilities to be used to implement
sync CRC algorithm negotiation in an upcoming change.