Add conversations_add_folder()
Which adds a new type of record to the conversations db, a "B" record
whose key is "B" + a string encoded CID, and whose value is a list of
mbox names separated by NUL characters.  These records will later allow
us to figure out which mailboxes need updating when CIDs are renamed.

Also upgrade conversations_dump() to print the new type of record.