Add conversations_rename_cid()
and conversations_rename_cid_mb() which is a convenience wrapper.  These
functions trawl through the conversations db renaming one CID to
another, which needs to be done when CIDs clash due to out-of-order
delivery on two masters or due to conversation merging.  The functions
also rename the "B" record for the CID and call a given callback for
each mbox named in the B record.  This callback will later be hooked up
to code in mailbox.c which will cause CIDs in the mailbox's index to be
rewritten at the next mailbox open.