First cut at XCONVFETCH
Separate out fetch-items parsing code into a new function so that it can
be shared between the FETCH and XCONVFETCH commands.  Add new
fetch-items UIDVALIDITY and FOLDER.  Add the command XCONVFETCH, which
given a CID and some fetch-items does a fetch for all the messages in a
conversation.

XCONVFETCH uses a "convstamp" technique to attempt to limit the amount
of data passed over the wire in the common case that the conversation is
unchanged since the last time the IMAP client checked.  The convstamp is
an opaque string to the IMAP client (actually implemented as the SHA1 of
the tuple (mboxname, uidvalidity,uidnext,highestmodseq) for each mailbox
touched by the conversation, in mboxname order), which has the property
that it is stable and does not change if none of the mailboxes touched
by the conversation have changed.  However the converse is not true: if
the convstamp changes, the conversation might or might not have changed.
The IMAP client remembers the convstamp returned in the tagged reply to
an XCONVFETCH, and passes it as an argument to the next one; the IMAP
server recalculates the convstamp from current data and if the convstamp
has not changed the server will emit no data.

Notes:

XCONVFETCH works only in SELECTED state, as it needs the currently
selected mailbox to implicitly choose which conversations database to
consult.

In this cut the untagged replies from XCONVFETCH are naked FETCH
replies, complete with the per-mailbox message number.  This needs to be
fixed.

Some extended fetch-item syntax should be disabled for XCONVFETCH but
isn't, like CHANGEDSINCE.

The FOLDER fetch-item reports the internal mboxname, which is wrong.