Next steps in evolution of XCONVFETCH
Remove the folder state stamp and instead pass complete folder validity
information as a list of (foldername uidvalidity highestmodseq) tuples.
When a change is detected, return the current state for each changed
folder back to the client using a new FOLDERSTATE untagged response.
This includes new uidvalidity and highestmodseq information and also a
VANISHED list for when messages are deleted.