Add POSITION, ANCHOR, OFFSET window args
 - POSITION specifies the 1-based index into the space defined
   by the searchargs and sortargs, of the first message to be
   reported.

 - ANCHOR specifies the UID of a message which, together with
   OFFSET, defines which will be the first message reported.
   If the anchor message is present in the space defined by
   the searchargs and sortargs, the position of the first
   message reported will be the larger of 1 and the position
   of the anchor plus the value of OFFSET.

ANCHOR overrides POSITION, i.e. if both are specified and the ANCHOR is
matched the ANCHOR will determine the position of the window, otherwise
POSITION will.

Also, reorganised the code a little so that the old index_sort()
function from a couple of commits ago is reinstated and new codepaths
are contained in a new index_convsort() function.  This helps with
short-term manual testing.