Emit empty FOLDERSTATE for missing folders
In the case where the client calls XCONVFETCH with a validity tuple
containing a folder which either does not exist or exists but is not
recorded as being touched by the given conversation, make the server
respond with a FOLDERSTATE reply containing NIL uidvalidity and NIL
highestmodseq.  This should allow the client to detect some obscure
cases involving folders being renamed or removed by other IMAP clients.