Add UNTIL to XCONVSORT
The UNTIL window item changes the reported window such that the position
designated by ANCHOR/OFFSET or POSITION is at the end of the window
rather than the start.  UNTIL is orthogonal to OFFSET.

Added an untagged response [POSITION $pos$] to XCONVSORT, which is
emitted if any results were emitted and indicates the 1-based position
in the sort order of the reported window.  Removed the internal implicit
MODSEQ search parameter in the CHANGEDSINCE case, as this was making the
POSITION reports incorrect.

Also, changed the format of the CHANGED list.  When the CONVERSATIONS
window item is specified, CHANGED reports conversation ids, but
otherwise it reports UIDs.

Also refactored the code to use two passes instead of a single pass with
a lookbehind buffer.  This made the code slightly slower but a lot saner.