conversations: simplify accounting
On second thoughts, we're also going negative because we copy these values
out and then delete the key even though there are still records waiting
to be renamed away from this CID.  And when they are, they will decrease
the count, and then when it gets to zero we can nuke the record anyway.

So let's do that.