Some more conversations API rejigging
Move the responsibility for enforcing consistency rules back down into the
conversations.c layer.  In particular, added conversation_update() which takes
a conversation_t and applies some deltas for exists counts etc and a modseq to
both the per-conversation and per-folder data in the conversation, enforcing
consistency.

Rename:

 - conversations_get_data -> conversation_load
 - conversations_set_data -> conversation_save

Add conversation_find_folder() which makes formal the find_folder() code the
tests were using.

Conversation_load() once again returns a NULL conversation if not found in the
DB.  The test code relies on this semantic to tell whether or not the database
state is correct.  Instead, change the 2 callers to handle that case.

Add a "dirty" flag in conversation_t.  This flag is clear on successful load,
is set on _new(), and is cleared on successful save.  Saves return early
success if the flag is not set.

Fix unit tests to mostly run again after the last few days' upheavals.  The
rename tests still fail and it's not clear to me whether the changed semantics
of the API are correct, so I'm letting them fail for the time being.

Conflicts:

	imap/conversations.c