unit: add config handling in mboxname tests
Allow the mboxname tests to fiddle with a limited subset of the global
config variables, while still providing isolation with other tests.
This is really ugly and hacky given the design of Cyrus' config code.

Also, fix the mboxname.nextmodseq test which broke a few days ago when
the code under test began behaving differently depending on the
"conversations" config option.