[IRIS-275] Fix XCONVSORT on self-expunged messages
XCONVSORT CHANGEDSINCE finds messages expunged in the same session and
reports them in the REMOVED untagged response.

Messages expunged in another session were working just fine.  Testing so
far has used a separate session (i.e. mutt in another window) to delete
& expunge messages, so didn't tickle this case.

The fix is rather gnarly and involves forcing a re-read of the entire
cyrus.index into a temporary index_state when CHANGEDSINCE is used.
There's got to be a better way.