xconvsort: regularise sort args
Having CONVERSATIONS be a singleton and everything else take an
argument is bogus.  Sorry.  I should have noticed this earlier,
but I didn't read the spec close enough.

This was causing bugs because of this even shittier thing up in
the Perl layer:

$Window = "(" . join(" ", grep { defined $_ } %$Window) . ")" if ref $Window;

Which was done because of this:

my $XConvSortArgs = { conversations => undef };

Ouch, ouch, OUCH.

And here's the end result when "position" wound up undefined.

<1301481996<45 xconvsort (reverse arrival) (conversations position 1 limit 30 offset -30) utf-8 all
<1301481997<4 xconvsort (reverse arrival) (anchor 5211 conversations position limit 30 offset -15) utf-8 all

IMAP is irregular enough already without making it worse.  The new format will look like this:

<1301481996<45 xconvsort (reverse arrival) (conversations 1 position 1 limit 30 offset -30) utf-8 all
<1301481997<4 xconvsort (reverse arrival) (anchor 5211 conversations 1 limit 30 offset -15) utf-8 all