annotate: precise quoting in ANNOTATION responses
In draft-daboo-imap-annotatemore-05 (which seems to be what we're aiming
at with the ANNOTATEMORE support), the mbox name is an astring but the
entry and attrib names are string, i.e. they must be quoted or literals
and cannot be atoms.  So, ensure this happens correctly.

Also, clean up some loop code.