Add mailbox actions file.
This is a file where actions can be posted, that are then later executed
by the first process to open the mailbox with an exclusive lock.  A
fcntl lock is used to provide exclusion between writers to file file,
and to detect when a writer has lost a race with an opener (who renames
the actions file before excuting the actions therein).  Actions are
simple lines of text.

Currently only a single action is implemented, the NOP action which just
increments a count.  This is for testing purposes.

The purpose of the mechanism is to allow delayed renaming of CIDs in
mailbox index files.

Merge summary:

    manually tweaked to add #include "strarray.h"