Remove message_update_conversations_file()
which was the wrapper for message_update_conversations() which parsed
the whole message into a struct body, for the two callsites which needed
to update conversations but didn't have a struct body.  One of these was
the APPEND path and the other was ctl_conversationsdb -b.  As it
happens, both of those have cyrus.cache information available so we can
just call message_update_conversations() with a NULL body*.