Use qsort for SORT,THREAD,XCONVSORT,XCONVUPDATES
instead of lsort - experiment indicates it's about twice as fast for
large mailboxes.  From callgrind reports of percentage distribution to
callees of cpu cycles, before:

index_convupdates
	30.84 lsort		<----
	28.00 get_modseq_of
	12.94 conversation_id_encode
	 9.16 hash_lookup
	 5.30 hash_insert
	 4.97 index_search_evaluate
	 3.01 index_msgdata_free
	 2.30 index_msgdata_load

and after:

index_convupdates
	32.07 get_modseq_of
	20.46 qsort		<----
	14.82 conversation_id_encode
	10.49 hash_lookup
	 6.07 hash_insert
	 5.69 index_search_evaluate
	 3.45 index_msgdata_free
	 1.70 free_hash_table

In absolute terms, lsort used 219,619,238 of however callgrind counts
cycles and qsort 127,211,813.  This was for a mailbox of 110,654
messages.