Predict correct TOTAL for XCONVUPDATES...ALL
when any expunged messages are present in the mailbox.  It turns out
that the index_state->want_expunged hack had the unfortunate side effect
of subtly changing the semantics of index_state->exists.  Worked around
by adding index_state->num_expunged which is updated in index_refresh()
but only if ->want_expunged is true (i.e. only for XCONVUPDATES).