[IRIS-1076] Another crash on empty folder
Index_msgdata_free() couldn't handle the case of a NULL msgdata* but a
non-zero count, which could happen in XCONVSORT and XCONVUPDATES if the
search total predictor reported a zero result (e.g. an empty folder).

Also, for XCONVUPDATES, we can't short circuit the main loop if the
mailbox currently has no messages but there are expunged messages which
might need to be reported as REMOVED.