[IRIS-1109] Multiple folders for GETMETADATA
Add a Cyrus-specific nonstandard extension of the GETMETADATA syntax
which allows multiple folder names to be specified in (parentheses)
instead of a single folder.  The folder names are not treated as
wildcards.

To support this, renamed annotate_apply_mailboxes() to
annotate_apply_mailbox_pattern() and add annotate_apply_mailbox_array()
which takes a strarray_t of mbox names.