Add mailbox_get_xconvmodseq()
and mailbox_update_xconvmodseq().  These functions pretend that some of
the state stored in the conversations database is a property of the
mailbox (logically, it really is).  They require the mailbox to have
been opened either writably or with MBFLAG_CONVERSATIONS, and return
IMAP_INTERNAL if this is not the case.