conversations: add conversations_set_suffix()
which allows the caller to change the way in which the pathnames of
conversations dbs are created from mboxnames.  Normally this is not
useful, but for conversation db audit we need to make sure that existing
mailbox code will do things into a special temporary db instead of the
live one.