[IRIS-1287] conversations: replace audit mode
Rewrote the code that implements ctl_conversationdb -A.  Now it:

 *  copies the live conversations db to a temp copy
 *  does a wipe & recalculate on the temp copy
 *  opens both live and temp copies
 *  scans through them comparing records and complaining if they differ
 *  if there were differences, tells you on stdout
 *  if -v is specified, tells you the differences in detail
 *  unlinks the temp copy

Note that we deliberately keep a lock on the user for the shortest
possible time.  The upside is that normal Cyrus operation is not held
out for long.  The downside is that the audit is racy w.r.t. new
messages arriving in the user's mailboxes, which may trigger a spurious
report of inconsistency.