conversations: sort by folder_number, not folder_name
We do this by not wiping the folder number on rebuild.

This allows us to have reliable reuse of folder names.  It does mean
that if you create and delete thousands of folders, you will get a
giant record - but that's life.

On the plus side, it means that B records should actually be BYTE
identical, not just "parsed foldernames" identical.