util: add buf_init_mmap()
which allows code that mmap() a region to store the result in a struct
buf which can then be buf_free()d later with the expected result of
munmap()ing the data.  This should help us simplify a few interfaces
in the message handling code.

v2: merged
e758e12 fixup! util: add buf_init_mmap()