Implement Sphinx search daemon start/stop
All the sphinx files, including the index, the config file, and the log
file, are in a per-user directory.  A new type of partition is
introduced, so we now have data, meta, and sphinx paritions.

A new imapd.conf entry, "sphinxpartition-$partition" defines the
directory which will be the top of the sphinx hierarchy.  For us this
will sit on large slow spinning rust as a sibling of the data partition,
e.g.

    /mnt/i18d12/sloti18d1p2/store122/sphinx/

The sphinx directories for each user are located under there, hashed
by user and (where provided) by domain, e.g.

    /mnt/i18d12/sloti18d1p2/store122/sphinx/domain/f/fastmail.fm/g/user/gnb/

or

    /mnt/i18d12/sloti18d1p2/store122/sphinx/a/user/aaallantrader/

The UNIX domain socket on which the Sphinx indexer daemon is listening,
is also per-user and is located in the normal Cyrus socket directory,
e.g.

    /mnt/ssd18/sloti18d1p2/store122/conf/socket/sphinx.gnb@fastmail.fm

or

    /mnt/ssd18/sloti18d1p2/store122/conf/socket/sphinx.aaallantrader

and this is cleaned up when the daemon is stopped.

Starting the daemon is done by running

    squatter -c start user.username

and stopping it likewise is

    squatter -c stop user.username

Starting creates the sphinx hierarchy for the user and also creates the
log file.  Starting is currently entirely manually initiated, a later
commit will make that automatic.

This was originally going to be a short-term expedient for automated
testing, but it looks like we're going to need it in production at
least for a little while.

v2:
merged
b36ec74 Add {start,stop}_daemon methods to search_engine
1e52c3c Add -c option to squatter for daemon control ops [some parts]
abd897e Fix various compile warnings [some parts]
3183acd fix compile warning in search_sphinx.c
eaaa39f sphinx: fix logic error when creating config file
e806588 sphinx: fix the 2nd start of the searchd
0594539 Sphinx files and socket are per-user
d6614bd search: add -e option to squatter [some parts]
794728e sphinx: fix compile warnings
c55cc67 Configure sphinx to log to syslog
30b36fd sphinx: configure syslog prefix
0c4974a sphinx: configure sphinx syslog better
2e15d8e sphinx: stop with --syslog-prefix too