sync: remove intermediate string for MD5 of records
and instead MD5 the values in the index_record as directly as possible
while preserving the algorithm.  We can make this kind of shortcut now
that various of the unuseful sync CRC combinations have been removed.
Further optimisations can be made but will require eithe breaking the
algorithm or changing struct mailbox.