[IRIS-1110] Add XWARMUP command
Syntax is, in ABNF

command =/ 'XWARMUP' SP mboxname SP '(' *warmup-items ')

warmup-item = 'INDEX' / 'CONVERSATIONS' / 'ANNOTATIONS'
		/ 'FOLDERSTATUS' / 'ALL'

The command does works whether or not a mailbox is currently SELECTed.
The command requires read permission for the mailbox.  The only result
are OK/NO/BAD, there are no untagged responses.  The command is not
announced in any capability.