[IRIS-1110] Add XWARMUP command
Syntax is, in ABNF

command =/ 'XWARMUP' SP mboxname SP '(' *warmup-items ')

warmup-item = 'INDEX' / 'CONVERSATIONS' / 'ANNOTATIONS'
		/ 'FOLDERSTATUS' / 'ALL'

The command does works whether or not a mailbox is currently SELECTed.
The command requires read permission for the mailbox.  The only result
are OK/NO/BAD, there are no untagged responses.  The command is not
announced in any capability.

[IRIS-1110] use mbentry for XWARMUP

Thanks to code review from RobM, who pointed out that index_open() is
going to read the entire cyrus.index file in blocking mode, which is not
quite what we want to achieve.

Suppress ENOENT errors in index_warmup()

Obviously, it's not an error to be unable to preload a file because it
doesn't exist.  But for some reason we've seen an ENOENT be reported to
syslog in production, so suppress it just before the error message
instead of where I thought it would be returned from a syscall.

[IRIS-1816] Handle posix_fadvise errors right

this time for sure.

[IRIS-1816] Ignore EINVAL from posix_fadvise

why did I ever check the return value.  Why???

[IRIS-1816] Suppress warmup errors properly.

Mondays.  I never could get the hang of Mondays.