Suppress ENOENT errors in index_warmup()
Obviously, it's not an error to be unable to preload a file because it
doesn't exist.  But for some reason we've seen an ENOENT be reported to
syslog in production, so suppress it just before the error message
instead of where I thought it would be returned from a syscall.