search: implement XCONVMULTISORT
which is a multiple-folder variant of XCONVSORT.  The implementation
varied from the specification in a few places, so read this carefully.

XCONVMULTISORT has the same syntax as XCONVSORT, except that the ANCHOR
window-arg is not accepted (because I couldn't figure out sensible
semantics for it).

Like XCONVSORT, it requires a selected mailbox.  There are two reasons
for this:

a) the search program parsing code fails if there isn't a selected
   mailbox and I couldn't be bothered fixing it, and

b) it's a handy way to let the server know which search scope (i.e.
   which user's search index) we want to look up.

The data is returned in an XCONVMULTI untagged response.  Why not
XCONVMULTISORT?  Because there's a horrible hack in Mail::IMAPTalk
which matches responses against /sort/i and tries to parse them as a
sequence of UIDs, which breaks the tests.

The format of the XCONVMULTI untagged response is

* XCONVMULTI (folder0 folder1...) ((fi uid) (fi uid)...)

where 'folder0' etc are external mboxnames, 'fi' are 0-based indexes
into the folders list, and uids are obvious.