sphinx: don't start searchd for trivial searches
so e.g. a search of "ALL" won't start the searchd.  We do this by
delaying the get_connection() call until the query is fully built, at
which point we can detect whether it's trivial or not.

This is useful optimisation, but also in the short term it allows
sphinx to be enabled on store254 without triggering pingsocket.