Add xstats: cheap & nasty metrics for Cyrus search
New command XSTATS reports a number of 32b counter metrics in an
untagged response.  Various points in the search stack are instrumented.
This is basically just a nasty hack to be able to measure various
performance-related parameters of the search code from Cassandane
without actually getting Cyrus' statistics working again.