squatter: better management of retry queue
Instead of keeping mailboxes on the retry priority queue forever,
complain to syslog and forget the item if we haven't been able to index
it after 10 minutes of delays.  Also, bump down the maximum retry period
to 1 second so that we get a decent number of chances to actually succeed.