squatter: unlink sync log after first pass
when all the mailboxes mentioned in it have been tried once, and not
when the queue is drained (which might be minutes away depending on
retries and exponential backoff).  This is more complex but ensures
that

a) we don't unlink the sync log file until we've at least tried to
   open each mailbox in it - the cassandane tests rely on this
   behaviour to synchronise.

b) however, neither do we delay unlinking in the pathological
   cases of long retries.  Instead we go around and give some
   other recently changed mailboxes a chance.

This should give us some measure of indexing fairness between mailboxes.