sphinx: per-user indexing lock
This puts a mutex around the lowest level of indexing operations that
need to be consistent, namely a single batch of messages (up to 20
messages).  Accesses and updates to the "latest" table are locked, as is
the code which calculates the next document id.  This should eliminate
the "duplicate id" message seen from sphinx in production.

Tested manually.