Replace mailbox_open_iwlnb with _irlnb
This was causing problems in production.  The rolling squatter could not
open a mailbox index for indexing because an imapd had the mailbox
selected and thus held the mboxname lock in shared mode.  Squatter would
retry many times with the same result and eventually give up and evict
the mailbox from its queue.  It turns out that squatter was opening the
mailbox in nonblocking but exclusive mode, which would of course always
fail.  The solution is to open the mailbox in nonblocking but shared
mode.  As squatter was the only code to call the mailbox_open_iwlnb()
API we just replace it with mailbox_open_irlnb().