sphinxmgr: account for started children correctly
While an indexd_t is in STOPPING state, it remains attached to the
hashtable and LRU list.  Thus the count of hashtable entries is not
an accurate measure of the number of started children.  Instead,
keep a global count which is incremented and decremented as index_ts
enter and leave the STARTED state.