sphinx: separate the two receiver classes
There are effectively two subclasses of search_text_receiver_t used for
the Sphinx integration code.  One is used when updating the index with
new messages, and the other in the XSNIPPETS command when generating
snippets for search matches.  Because both of them talked to searchd via
the MySQL protocol and both of them received message text, they were
implemented using the same struct.  However we want to reimplement the
way updating is done, while leaving the snippets code alone, so as a
first step we separate the two into different structs.