sphinx: call {end,begin}_mailbox when yielding
One important side effect of the previous indexing batching code was
that the search engine was notified of the boundaries of batches, so
that it can actually flush whatever it might have pending when a batch
ends.  This effect was lost in the shuffle when this code was refactored
to yield the mailbox lock rather than go back through the mailbox close
and open path again.