xapian: set the stemming strategy during indexing
This uses a feature of Xapian 1.2.11 and above which allows the caller
to control how stemmed words are indexed.  We choose to stem all words
and save only the stemmed variant, without a Z index.  This should
minimise the disk space requirements while still giving us stemmed
searches and positional information (which will be needed for snippets).