index: minimise the size of struct index_map
This reduces the memory usage for a large index considerably,
from 256 bytes per record down to just 48 bytes.  That's a
pretty massive saving for a big mailbox!

This is achieved by only keeping copies of the mutable data about
each record in the index_map.

There are two new functions - one for re-reading the record and
applying the mutable data to it, and another for writing a record
and updating the mutable data in the map to match.

Because the re-reads may be done without a cyrus.index lock held,
they can contain incorrect data in the mutable fields, but this
gets overwritten by the snapshotted version in the map before use.