Adjust search_expr:split unit tests
so that the input expressions are properly normalised.  The current code
expects normalised expressions but certain details of normalisation
don't matter, e.g. the order of child nodes.  This will soon change so
that the splitting code fully normalises it's input, which will affect
the order of some outputs from the splitting, which spuriously breaks
these tests.