Skip search expr normalisation if possible
The purpose of normalisation is to allow splitting by indexable
attribute and by folder to make index lookup and folder scanning as
efficient as possible.  If the expression contains neither of those
kinds of nodes then there's no value in normalisation; but there is
potentially an exponential explosion in memory and CPU cost for some
cases involving far too many OR nodes.  So we skip normalisation if
the expression contain no folder match nodes and no fuzzy match nodes.