index: track generation in index state
This can never change due to the held namelock at the moment, but
later it will be possible to release the namelock and allow a
repack.

We then need to know the repack has happened so we can sweep through
and update the recno pointers to their new locations in the
cyrus.index file.  Luckily, the generation_no header field already
tracks exactly this.

(NOTE: repack can happen without bumping highestmodseq, in fact it
 usually does, hence the need for a separate check)