cyrusdb: add cyrusdb_lockopen()
This is exactly like cyrusdb_open, but you pass a 'struct txn **'
and get back a database already locked.

This commit just implements it in terms of a call to the backend's
lock code after opening, but in future at least some backends can
optimise by knowing that the caller wants a writelock, and avoid
releasing and re-obtaining the lock.