xapian: handle locked indexes sanely
We may have multiple writers trying to write to the same index
at the same time.  If you try to open a Xapian index and it's
already locked, there's no option to wait for the lock to be
available, you just get a "Locked" exception.

NOTE: it's still possible (small race window) to double-index
the same messages with concurrent writers, but I'm not too
worried about that cost, and it will be mostly fixed by switching
to SHA1 based indexing later anyway, where the record is separately
looked up in the locked Xapian index before re-indexing.